+++ db/postgresql/update-3.sql
@@ -1,0 +1,3 @@
+-- Add expiration on account requests (and accounts)
+
+ALTER TABLE idp_account_request ADD COLUMN expiration timestamp DEFAULT NULL;