+++ db/postgresql/update-4.sql
@@ -1,0 +1,2 @@
+-- Email is now optional for tutorial accounts
+ALTER TABLE idp_account_request ALTER COLUMN email DROP NOT NULL;